# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Subscriptions Aggregation API
  version: 1.0.0
servers:
  - url: http://localhost:8080
paths:
  /health:
    get:
      summary: Проверка состояния сервиса
      description: Возвращает статус `ok`, если HTTP-сервис работает.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
  /api/v1/subscriptions:
    post:
      summary: Создать подписку
      description: Создает новую подписку для пользователя.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SubscriptionRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Subscription' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '500': { $ref: '#/components/responses/InternalError' }
    get:
      summary: Получить список подписок
      description: Возвращает подписки с фильтрацией по пользователю и сервису, а также пагинацией.
      parameters:
        - in: query
          name: user_id
          schema: { type: string, format: uuid }
        - in: query
          name: service_name
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, default: 50 }
        - in: query
          name: offset
          schema: { type: integer, default: 0 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Subscription' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/v1/subscriptions/{id}:
    get:
      summary: Получить подписку по ID
      parameters: [ { $ref: '#/components/parameters/ID' } ]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Subscription' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
    put:
      summary: Обновить подписку
      parameters: [ { $ref: '#/components/parameters/ID' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SubscriptionRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Subscription' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
    delete:
      summary: Удалить подписку
      parameters: [ { $ref: '#/components/parameters/ID' } ]
      responses:
        '204': { description: Deleted }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/v1/subscriptions/total-cost:
    get:
      summary: Подсчитать суммарную стоимость подписок за период
      description: Считает сумму по месяцам, пересекающимся с указанным периодом.
      parameters:
        - in: query
          name: from
          required: true
          schema: { type: string, example: '07-2025' }
        - in: query
          name: to
          required: true
          schema: { type: string, example: '12-2025' }
        - in: query
          name: user_id
          schema: { type: string, format: uuid }
        - in: query
          name: service_name
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  total_cost: { type: integer, example: 2400 }
        '400': { $ref: '#/components/responses/BadRequest' }
        '500': { $ref: '#/components/responses/InternalError' }
components:
  parameters:
    ID:
      in: path
      name: id
      required: true
      schema: { type: string, format: uuid }
  schemas:
    SubscriptionRequest:
      type: object
      required: [service_name, price, user_id, start_date]
      properties:
        service_name: { type: string, example: Yandex Plus }
        price: { type: integer, minimum: 1, example: 400 }
        user_id: { type: string, format: uuid, example: 60601fee-2bf1-4721-ae6f-7636e79a0cba }
        start_date: { type: string, example: '07-2025' }
        end_date: { type: string, nullable: true, example: '12-2025' }
    Subscription:
      allOf:
        - $ref: '#/components/schemas/SubscriptionRequest'
        - type: object
          properties:
            id: { type: string, format: uuid }
            created_at: { type: string, format: date-time }
            updated_at: { type: string, format: date-time }
    Error:
      type: object
      properties:
        error: { type: string }
  responses:
    BadRequest:
      description: Некорректный запрос
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    NotFound:
      description: Запись не найдена
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    InternalError:
      description: Внутренняя ошибка сервера
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }